refactor(workflow-designer): remove vanilla interface and simplify
