+++ packages/workflow-designer/src/hooks/use-properties-panel.ts
