+++ packages/workflow-designer/src/index.ts
@@ -1,1 +1,6 @@
-export { WorkflowDesigner } from "./workflow-designer";
+"use client";
+
+export * from "./context";
+export * from "./hooks";
+export * from "./utils";
+export * from "./types";
