--- packages/workflow-designer/src/react/index.ts
@@ -1,4 +1,0 @@
-"use client";
-
-export * from "./workflow-designer-context";
-export { useWorkflowDesigner, ViewState } from "./state";
