+++ packages/workflow-designer/src/utils/index.ts
@@ -1,0 +1,1 @@
+export * from "./is-supported-connection";
