+++ packages/workflow-designer/src/utils/is-supported-connection.ts
